make sret a long by default.  strlcpy and strlcat return size_t (unsigned long), and snprintf returns int.  we want to use one function to check the return value of all three string functions.

+++ import/import_af6.c
@@ -49,7 +49,7 @@ FILE *vfd, *afd;
 
 MOD_open
 {
-    int sret;
+    long sret;
 
     if(param->flag == TC_AUDIO) {
